Use size_t in function signature

+++ src/boosting/gbdt.cpp
@@ -249,18 +249,19 @@ void GBDT::Train(int snapshot_freq, const std::string& model_output_path) {
   }
 }
 
-void GBDT::RefitTree(const int* tree_leaf_prediction, const int nrow, const int ncol) {
+void GBDT::RefitTree(const int* tree_leaf_prediction, const size_t nrow, const size_t ncol) {
   CHECK_GT(nrow * ncol, 0);
-  CHECK_EQ(num_data_, nrow);
-  CHECK_EQ(static_cast<int>(models_.size()), ncol);
+  CHECK_EQ(static_cast<size_t>(num_data_), nrow);
+  CHECK_EQ(models_.size(), ncol);
+
   int num_iterations = static_cast<int>(models_.size() / num_tree_per_iteration_);
   std::vector<int> leaf_pred(num_data_);
   if (linear_tree_) {
     std::vector<int> max_leaves_by_thread = std::vector<int>(OMP_NUM_THREADS(), 0);
     #pragma omp parallel for num_threads(OMP_NUM_THREADS()) schedule(static)
-    for (int i = 0; i < nrow; ++i) {
+    for (size_t i = 0; i < nrow; ++i) {
       int tid = omp_get_thread_num();
-      for (size_t j = 0; j < static_cast<size_t>(ncol); ++j) {
+      for (size_t j = 0; j < ncol; ++j) {
         max_leaves_by_thread[tid] = std::max(max_leaves_by_thread[tid], tree_leaf_prediction[i * ncol + j]);
       }
     }
@@ -268,6 +269,7 @@ void GBDT::RefitTree(const int* tree_leaf_prediction, const int nrow, const int 
     max_leaves += 1;
     tree_learner_->InitLinear(train_data_, max_leaves);
   }
+
   for (int iter = 0; iter < num_iterations; ++iter) {
     Boosting();
     for (int tree_id = 0; tree_id < num_tree_per_iteration_; ++tree_id) {
